ci(docker): change how we tag images to ${REVISION}-${SHA}
This change:

* builds images tagged ":tmp"
* upload them as :${REVISION}-${SHA}
* retag them to :latest
* release :latest with heroku container:release

This will allow to retrieve image from Heroku and revert to an old one
if needed, it also have the benefit to put SHA/REVISION inside the docker image.

Pushing and releasing the "latest" tags should be atomic in a perfect
world. As this not possible, this 4 seperate steps ensure we don't have
docker image layers upload during the critical section, just tagging and
release.