feat(post_check): allow to set success_conditions (#277)
Since we move base attribute to editable attribute, this have the side
effect to always post the post_check check-runs whenever the base match
or not.

We cannot just revert the previous change as when base attribute is
changed on a pull request it was triggering a lot of bug, like
post-check not been reevaluated or cancelled, the merge-queue may got
stuck because cancel() is never called, ...

So instead this change offers more flexibility to the action. User can
now set a different set of condition to make distint which conditions
make the check-runs appear on the pull request and which conditions mark
it as failed or succeed.

Fixes MRGFY-1300