fix(dismiss_reviews): filter out too recent reviews (#369)
If an user request a review, we receive an event and start processing
the pull request, we got the pull request detail with
`requested_reviews` set, then the requested user approves the pull
request while mergify is processing the PR, then we retrieved the
reviews list and the approvals is there (and GitHub has cleaned
`requested_reviews` on its side.

Next the dismiss_reviews seen that an approvers is in the
`requested_reviews` and dismiss the review while it should not.

To work around the issue, we track the most recent timestamp of events
we receive and drop any review done after. This keeps coherant
 `requested_reviews` and the `reviews` list. So dismiss_reviews action
 will not take wrong decision.

Fixes MRGFY-1150