refactor(actions): evaluate action configuration earlier (#435)
This allows to evaluate action config against the current context
earlier in the process.

This will permits for action like post_check to compute some
configuration data (eg: success_conditions) against the context
states as the pull_request_rules conditions. So if an action
change/update the context post_check/success_conditions will not be
impacted by this (and later new events will retrigger the engine).

To do so, the Action class in split into 2 parts, the one that just
embed the raw configuration and the one that evaluate the raw config for
the current context and have the run/cancel method.

This also permits to compute the final state of a jinja template and
returns an error before the action is effectively ran.

Theses runtime errors will now be returned via the simulator.

Another benefit side effect of the split is that we need to type the
interpreted configuration of an action.

Fixes MRGFY-1356