ci(snyk): drop it (#603)
It's unusable for us, each time the base image change it creates a new
project in Snyk and we loose all acknowledged CVE and the reason. Synk
support tell us this how it works, new base image is a new project which
is not from our point of view. And they didn't propose any solutions.