refactor(worker): change when we compute the next PR run (#583)
Currently, we compute the next PR run when we try to remove the pr from
the org stream. If new events have been received in the meantime, we
pick the score of the next events. But since this event arrive in the
meantime, the score is already in the past so the pull request need to
be reproceed now.

This changes compute the next time to refresh a PR when we get all
events, so even if events arrive when we proceed the PR it will not
override this score and just stack more events in the stream.

The delay between two processes of the same PR has been set to 15s if
the worker is busy, nut if the worker has nothing to do it will process
them before.

Depends-On: #903
Depends-On: #906