fix(worker): retry MergeableStateUnknown faster (#1343)
When something got merge in main branch, the mergeable_state of pull
requests switch to unknown. We need to wait a bit before retrying.

A recent change has cap retry to 30s minimun, which looks good for
GitHub connectivity issue or increasing the number of events we proceed
on each run but for this particular case we should be faster.

Change-Id: Ib91a30c8be693b092986fe89110b83c5222e0f60