refactor(yaml): use C parser (#1375)
Here the result of different yaml Loader by parsing 100x the airtable config file:

* Loader: 3.743259792000572 seconds
* SafeLoader: 3.7517577500002517 seconds
* CSafeLoader: 0.2654330410005059 seconds
* CBaseLoader: 0.2304462500005684 seconds
* Dumper: 1.77089741600048
* SafeDumper: 1.764061999999285
* CSafeDumper: 0.2553510420002567
* CBaseDumper: was not able to dump the config

CBaseXXXXXx can't be used, our tests won't pass with it (some types are
not supported like boolean).

So we pick CSafeLoader/CSafeDumper.

This change have a drawback, the error reporting of the C parser is not
as good as the Python one.