revert: "chore(count_seats): also count for user triggering CIs (#1484)" (#1534)
This is actually a bad idea because come CI like CircleCI will use a
random user token to trigger a workflow, and this will be reported by
GitHub by being the user.

This can confuse customers when they have people not working but being
counted as active because their tokens is used.

This reverts commit fc4f9428dc36cb86e92322addf064d85d2f9b9bc.