refactor: drop usage of mergeable state (#1554)
Waiting for GitHub asynchronous operation to finish is buggy by design.
This makes the engine retries pull request processing a lot.

The trade of is that conflict check in merge/queue has been dropped, we
will now rely on GitHub merge API to return an error in such case.

Also, this change how the `conflict` attribute works has `mergeable`
could be None until GitHub computes its state. The None value will be
considered as "True" if we don't have the last value in cache. And when it's None we will refresh the pull
request later to see if it's