fix(config): configuration extends recursion (#1571)
Extending configuration doesn't work as expected, depending of how
we load the configuration, we don't have the same behavior.

And the recusion limit doesn't work, the associated tests was not
testing the recusion but only if the allow_extend param works.

This changes the default `allow_extend` to `True`, so anywhere we load
the configuration the extend will works as expected.

Also we set it to false when we load the extends configuration to avoid
infinite recusion.

And the tests are fixed.

Fixes MRGFY-1547