chore: remove useless parameter (#1702) (#1711)
Recent refactor has enforce 30s between pull requests.
This works well if we always have events for a pull requests.
But in worse case scenario if we receive an event just after we emptied
the last batch of events, we will reprocess the pull request instantly
without waiting the 30s. This is because the Redis ordered set entry has
vanish and we no longer has the score tell us when we proceed this PR
the last times.

This change add the 30s delay to the score to ensure we proceed PR only
each 30s if no events was there before to avoid this race.

Fixes MRGFY-1560

Change-Id: If2a204944b7234be5a9326ae6e3fbc80cceeac0b