fix(refresh): stop refreshing PR closed after 7 days (#1730)
If conditions has something like:
```
 - current-time>=8:00
 - current-time<=17:00
```

This is going to refresh all pull requests opened and closed for ever.
The backlog of pull requests we have to refresh at 8h00, 17h00 and
00h00 can only increase.

This change stop refresh closed PR after 7 days. We keep 7 days to keep
the summary update to date in case user investigate a recently merged
PR.