chore(rules): avoid useless partial dict copy (#1712)
For an unknown reason, the code does a non-deep copy of some dicts,
return the non-deep copy, but modify the original one.

This was working only because the values of each dict was shared.

This changes stop doing the useless non-deep copy and stop returning
dicts to make clear each function updates the dict passed to it
and does not copy anything.