fix(queue): unqueue PRs when configuration change (#1706)
Changes on Mergify configuration can cause a `RuntimeError` for pull requests already queued. Before adding a PR to the queue, we check that every PR in the queued is synced with the configuration. Otherwise they are removed from the queue to be re-embarked later with the new configuration.

Fixes MERGIFY-ENGINE-2TS