fix(eventlog): add UnexpectedChange PR removal report to EventLogs (#1494)
when a PR wa removed from the queue or the train was reset due to unexpected changes, 
we couldn't see the exact reason why in the eventlog. What we saw instead was the generic
report being shown in the "Rule :" summary, which does not give any detailed information
about what has really happened. Now we can display the proper report following
previous changes from the refactor of TrainCarInformation datastructure

Fixes MRGFY-1460