fix(queue): add train reset when UnexpectedBaseBranchChange while merging batch (#2415)
While a batch was being merged, if a base branch change occured, we did not reset the train, thus letting the batch continue its merge. Due to latency during a past week, this bug occured and unveiled the problem. Now, we test if the branch train in properly synced with the base branch before continuing the merge.

Fixes MRGFY-1540