fix(queue/merge): retry head/base branch failure (#2589)
When merge failed for transient reason, the queue may be stuck as the
action return pending and wait for the next event to refresh the PR.

For base branch change event for example, we may receive the push event
a lot of time later because this is a low priority event.

This change sends a immediate refresh to reevaluate the situation.