fix: retry config check when PR mergeable state is unknown (#2755)
Sometimes GitHub may not have yet computed the merge_commit_sha and it's
None. This makes the configuration checker to download the configuration
file of the main branch instead of the one from the pull request.

Then depending on what merged recently, we can miss a configuration
change or wrongly report the configuration has changed.

This change retries the PR later in such cases.

Fixes MRGFY-1667
Fixes MRGFY-1534
Fixes MRGFY-1728