ci(poetry): pip bug workaround (#2794)
pip has a bug when it build pep517 wheel, the isolate virtualenv to
build the wheel is not really isolated, pip/setuptools libs are just links to
the poetry/tox virtualenv libs. So when peotry upgrade setuptools, some
files could vanish for a short times.

Poetry 1.2.3 or 1.3.0 will workaround automatically that with:
https://github.com/python-poetry/poetry/pull/6582

In the mean time we update the tox virtualenv ourself.

Fixes MRGFY-1711