fix(date): Fix `Schedule::get_next_datetime` (#2855)
The conditions at the beginning of the conditions were out of place,
this caused the function to return a datetime out of schedule when the
`from_time` was already out of schedule and outside time limits.

Also, the return value if we were inside Schedule was pointless. We
returned a value at end_hour and end_minute, but it is useless since we
have other way of checking if a PR can be merged or if an action can be
done before a Schedule expires.