fix(tests): explicitly close postgres connections (#2927)
asyncpg connections are tied to a asyncio loop.

When the test finishes, the APP_STATE is set to None, the engine need to
be garbage collected, when the garbage collector calls __del__ on the
engine, this one close all connections and if the asyncio loop has been
closed this fail.

Most of the time we are lucky because of one of these reasons:

* The connections are closed and the engine has been disposed by the garbage collector before pytest-asyncio create a new loop.
* The garbage collector is never run before pytest ends

This change closes the postgres connections explicitly and dispose the
engine manually to ensure the garbage collector will not have to do it.

Fixes MRGFY-1712