refactor(queue): factorize the train car mergeability state (#3150)
Recent code changes like the precise train car outcome, raw dict of
last_conditions_evaluation, the fixes about unexpected change detection
have reduced the difference when we do the mergeability check of a train
car between a draft pr and an in place check.

This change refactors the mergeability check to do it in only one place
"TrainCar.check_mergeability()". And call this new helpers in 3 keys
parts. This change also remove the complex refresh loop between a draft
PR and the user PRs.