chore(squash/rebase): deprecate random user pick (#3368)
This adds a notice to user each time we rebase/squash a pull request
with a random user. We don't spam them if this is done by the author as
this will be the default behavior later.

The new failure message also fake the new behavior to force author not
logged in Mergify to do it in case of no token match.

Fixes MRGFY-1838