chore(queue): better unqueue messages (#3015)
This change makes mandatory to set an unqueue reason to `remove_pull()`.

It's now up to the queue action to tell why it removes the pull request from
the queue. The action will choice if this is a PrDequeued due to
pull_request_rules, configuration issue, etc.. or is this is due to TrainCar
outcome.

This introduces new unqueue/abort reason like:

PR_MERGED
PR_DEQUEUED
TARGET_BRANCH_MISSING
TARGET_BRANCH_CHANGED
PR_FAILED_TO_MERGE

For example, this change this wierd message like this:

`Checks aborted for pull request #2853: Pull request #2853 which was
ahead in the queue has been dequeued.`

To:

`Checks aborted for pull request #2853: Pull request #2853 has been
unqueue.`

Also, when the unqueue reason is due to the train car outcome this can now be
reported in the action summary.