fix(queue): avoid queue action become pending again (#3576)
The action_runner may rerun queue action for nothing. This fixes it.

But the fix showed that the action check-runs were not always good and were
fixed by a subsequent rerun.

So this also changes how we detect the head of the train since we knew we merged the PR ourselves now, and we could generate the right report on the first try.

The pull request can't be split as each change exposes the bug of the other change.

The two bugs were hiding from each other.

This reduces a lot the number of GitHub API calls done by queue action.

Fixes MRGFY-1852