fix(queue): event ordering race condition (#3745)
If during the evaluation a pull request in the first batch but not at the
first position, the last CI of the draft PR posts its results, the
evaluation marks the draft PR as successful and mergeable, but the first PR
in the queue is not refreshed to be merged.

This change how we detected we need to refresh the next pull request to
merge.

Depends-On: https://github.com/Mergifyio/engine/pull/3744
Fixes MRGFY-1869