fix(queue): priority_rules with priority < medium (#3896)
Currently, if the priority_rules mechanism returns a priority lower than
the priority set in the action configuration, the priority from the
action configuration is returned. Since the default of the action is
"medium" anything lower than "medium" returned by priority_rules was
overrided by the action default "medium".

This change the way we compute priority to the priority in action
configuration only if the priority_rules mechanism didn't match any
rules.