fix(backport/copy): don't reuse any rule/action objects (#3905)
We delete a couple of times branch we shouldn't delete because the
internal Mergify shouldn't have matched. Also we don't get the
associated logging when this occurs meaning the ctxt used for deleting
the branch with the delete_head_branch is not the same from context used
for logging.

Anyways, action objects are not design to be reused.
So this change remove the global.

Fixes MRGFY-1850