fix(queue): delay branch protection configuration check
Currently, we check the branch protection setting of target branch
against all queues configurations.
But if a queue target only one branch, we can raise incompatibility against
a branch configuration that will never be used (example: base=XXX will never
match)

Since we can't really guess in advance if in which queues a pull request
can go because of the conditions system.

We have no choice that controlling the configuration with we plan  to queue it (eg:
run/cancel()).

Fixes MRGFY-1917