fix(dismiss_reviews): timing issue (#4107)
If a review is done while Mergify is processing the synchronize event,
this review may be dismissed even if it has been done after.

This change makes leverages synchronize event received_at attribute to
compare it with the reviews.

It's not perfect as the received_at may not exactly when synchronize
happen, but it far better than nothing. The timing conditions to trigger
the bug may it almost not possible.

Fixes MRGFY-1552