fix(redis): use blocking connection pools for Redis client used by API (#4126)
These days we got some events missing because our API was reaching the
max_connections settings. It's not yet clear why we have a limitation
since we pass None to max_connections. We have new Datadog graph
"Connected clients by sources" to understand this:
https://app.datadoghq.com/dashboard/t5g-9mk-2pb/redis?live=true

Using unlimited non-blocking connection pool for third party services in
an API is bad practice as we don't control the how many requests can be
done in //.

This change uses a blocking connection pool instead. So if the
max_connections is reached, new HTTP request will wait for a while
instead of returning a 50X.

All max_connections are already set to higher values (200) in production.