fix(freeze): drop application_id/name from queue-freeze (#4308)
The resource should not be tied to an application. Since we migrate to
the engine backend we known which user click on freeze button.

What we really need is to feed the eventlog with who
create/update/delete the freeze. Where `who` can be an application or an
user.

Fixes MRGFY-1966