fix(copy/backport): always finish them (#4452)
In case we cancel the action on a half created backport, it's too risky
to cancel it and clean up created resources. People may have rules that
invalidate the condition on purpose after the backport has been created. We
should keep the same behavior as before, keep it.

So this change ensures we always finish the started change.