fix(worker): latency metric (#4576)
A bucket can contain sources that have different prio.

For example, if a pull request got a low-priority event, the bucket
priority/score will be low, then the pull request received an immediate
priority/score, the bucket priority/score change to immediate, but the
old and new events are inside the bucket to batch them efficiently.

This change keep track of the initial priority/score to build the latency
metric correctly.