fix(queue): always evaluate all pull requests of a train car (#4575)
Currently, the UI looks wierd as while we are merging pull request of a
batch one by one, the queue rules printed in the summary change.

This change avoids to change the queue rules.

And may avoids re-updating all check-runs each time a pull request of a
batch is merged.