fix(copy/backport): ensure we periodically refresh backport/copy state (#4633)
In case of copy/backport is interrupted, we should recover
automatically. This is already done when the PR received a new event and
the excepted is not there we recreate it.

So, to not rely on next received GitHub event to recover, this change
automatically refresh the PR every 3 minutes until the backport succeed
or fail.

Fixes MRGFY-1992