perf: faster shutdown (#4676)
We currently wait 10 seconds for the worker to shutdown properly
otherwise we kill all asyncio tasks.

But the worker loop check if the stopping condition is set every 30s.

So in production, workers are always killed except if we are lucky about
the timing.

Also during shutdown Datadog agent exit before worker. This result to
the lost of all shutdown traces/logs.

This change just relies on Cancellation mechanism to shutdown the
worker. If one day we really need to not interrupt a task, we must use
asyncio.shield() instead.