fix: change retry strategy of dashboard API calls (#4743)
The default retries strategy is too high for dashboard, up to 6s.

Since we have good chance to need the subscription and/or users
tokens, this slow down a lot the engine while we have the data in cache
and if not we still have the worker that will retry the event later.

This changes the maximum to 600ms for dashboard API calls

Fixes MRGFY-1980