fix: put different pg pool size per dyno (#4836)
We used 100, because only the web dyno was using postgres until this
morning. Now the workers also connect to postgres to retrieve tokens.

We have only 120 connections available and 4 workers running, so
we put 15 connections each, 50 for the web worker and keep the rest for
datadog and Heroku monitoring

Fixes INC-59