fix(bot-account): inform later that the configuration is invalid (#4867)
Since we dropped the random user pick for bot_account, we picked the author of the PR. But if this one
if a GitHub App (eg: dependabot[bot]), Mergify is currently fully blocked because the configuration is invalid.

This change instead will report the configuration issue when the action run. This permits not blocking action that doesn't have `bot_account`

Related [#5074](https://github.com/Mergifyio/mergify/issues/5074)