feat: create a Redis cache for commits (#4972)
Create a new Redis cache to store commit SHA and the pull requests
associated. This is used by the CI Optimizer dump script to be able to
link a job run with a pull request, even if the pull request's SHA has
changed.

We can't use the PullRequestFinder, because it doesn't store the SHA
history of each pull request.

Fixes MRGFY-2031