fix(queue): ensure head_sha and commits are in sync (#5098)
is_behind is not super reliable to detect if a pull request is up to date.

Here the new race condition, we use GitHub API to merge main branch in
the pull request and wait for is_behind to become false.
Unfortunately, when we detect the pull request is up to date (by looking
at the list of commits of the pull request), the head_sha of the pull
request may not have been updated by GitHub yet :(
We think the PR is up to date and gather the CIs attached to this head sha
of the pull request, everything looks green and we click on merge button.
But the head sha we have was an old one... the return an error and the
pull request is unqueued.