chore(marketplace): remove useless endpoint (#5131)
The current endpoint reset subscription and then forward the marketplace
event to the dashboard.

The dashboard just mark the accounts that need to be refreshed.

The reset of the subscription is useless as the engine will retrieve
instantly an outdated subscription from the dashboard. And later will
get the subscription reset again when synack will run.

This change removes this endpoint.

The marketplace event will be sent to the dashboard directly.