fix(rebase/update): Do not allow to rebase a PR in queue (#5250)
Currently, if the user rebases the PR with Mergify, the queue is not reset as we ignore our own push/synchronize.

If a draft is created, the checked stuff is different from what will be merged.

If branch protection strict is enabled, this will reset the queue when we try to merge it and the user doesn't understand it's because of its manual rebase.

So we need to handle this case by either:
1. just ignoring this rebase (as the queue is here to check that.)
2. removing/readd this PR from the queue.

This PR implements option 1.

Fixes MRGFY-2083