fix(queue): Change message when creating draft pr with base branch conflicts (#5753)
When creating a draft PR with a pull request that has conflicts with the
base branch, the message we displayed in the summary was the same
message as the one we use when a draft pr conflicts with pull requests
ahead in the queue. Except that in this case there are no pull requests
ahead in the queue.

Depends-On: #5742