test(db): allow only one head revision (#6043)
`test_migration` won't work if there are two head revision. I added a test to check that there is only one head revision. The assertion will help the developer to spot the two revisions conflicting.

The "two head revisions" happens when 2 developers create a revision with the same base revision.