refactor(ci): move JobRun constructor to its own class (#6257)
There is a JobRun constructor `create_job` that was tied to HTTPJobRegistry,
whereas it has actually now link to the HTTPJobRegistry per-say.

This made a very weird call being made to the HTTPJobRegistry class to create a
JobRun, which barely sense. This moves the JobRun class constructor to its own
class.