fix(ci): process all events (#6314)
We can't limit the processing to 5000 events. We try to process every
job run until we receive its workflow run event. It appears that a lot
of job runs ends long before the workflow completes. So we have to process
all the stream events for now and lower the expiration of events.

This solution is far from perfect. We'll work on a new design of the event
processing that will happen in several steps. We have to make it work for
now and monitor it, to implement the final solution incrementally.

Fixes MRGFY-2286