feat(queue): add Merge Queue Pause (#6481)
We wanted to be able to pause a merge queue for
emergency cases. The pause is cutting every current checks running, and prevent new ones from starting. PRs can still be added to the queue but they will not be checked. Once the pause is lifted, the Merge Queue resumes where it was.

Fixes MRGFY-2054