chore(db-migration): test migration script individualy (#6544)
alembic uses only one sqlalchemy Metadata object per upgrade and doesn't
cleanup it between script.

So if a metadata resource (like an enum) is used in a migration script,
next script will reuse it instead of creating it.

This breaks assumption of following scenario:
* script 1 create enum foo
* script 2 reuse enum foo

If 1 has already been applied on the database, but not 2, when we run
the script 2, alembic will not load the script 1 and will not known enum
so already exists.

This change applied all scripts one by one to ensure their work when
ran isolated.

Fixes MRGFY-2370