fix(queue): return CANCELLED instead of FAILURE (#6589)
FAILURE is a fatal state, the queue action will never ever run after
this.

If routing conditions or branch protection have an issue, we must return
CANCELLED, so when the issue is fixed, the queue action can put back the
PR in queue.

Related to MRGFY-2376, but not the root cause of the initial failure.