ci: fix create-linear-for-docs (#6731)
The job failed with:

{"errors":[{"message":"Syntax Error: Expected \"$\", found \":\".","locations":[{"line":1,"column":10}],"extensions":{"code":"GRAPHQL_PARSE_FAILED","type":"graphql error","userError":true,"meta":{}}}]}

The error message is indicating that the GraphQL query has a syntax
error. It expects "$" but found ":", meaning there might be a problem
with the variable definitions or JSON formatting. I realized there was
an error in the request body format in the previous examples.

The correct format to use variables in GraphQL is by declaring them
outside of the query, then use them inside the query with $ prefix.