ci: ignore new semgrep rule (#6846)
A new semgrep rule enforces to uses "sha" instead of tag for GitHub
action.

This is a good security rule, but when you switch from tag to sha you
don't get Dependabot updates anymore.

As GitHub state that if you trust the author you can use tag instead of sha:

https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions#using-third-party-actions

Here are the GitHub actions we decided to trust:

* Mergifyio/*
* docker
* aquasecurity/trivy-action

* hadolint/hadolint-action
* amannn/action-semantic-pull-request@v5
* marocchino/sticky-pull-request-comment
* peter-evans/create-pull-request