fix(worker): workaround for unexpeceted CancelledError (#6992)
This change is a bandaid to recover to any unexpected
asyncio.CancelledError.

Couple of bug in in the past has show this can occur and INC-78 proves
bug like that can arrive again.

This change avoids stopping an asyncio.Task that has not being explicly
stopped by us.

Related INC-78
Related ENGINE-3CK