test: fix gpg key (#7027)
We recently change mergify.io to mergify.com, but we miss the gpg key
because it was set in the environment variable instead of the settings
file.

This change put the key in the configuration file as it's not really
private as everybody has it in they configuration file.

I renamed the configuration settings to not load the old environment variable from developer config file