refactor(console-cli): introduce click (#7206)
We have a lot (15 console scripts), all of them have different
interfaces (argparse or direct argv parsing) and purpose (docker
entrypoint, dev tools, admin/support tools).

The plan is to have only new console scripts that rule them all. The
goals are to have only one way to prepare the cli, passing the arguments,
etc...

We will keep for docker images Saas and on premise:
* mergify-database-update
* mergify-engine-worker
* mergify-count-seats

We will introduce `mergify-admin` to replace:
* mergify-redis-dump
* mergify-stream-reschedule
* mergify-stream-status
* mergify-debug
* mergify-clear-token-cache
* mergify-refresher
* mergify-merge-queue-reset
* mergify-api-internal-openapi-spec
* mergify-api-public-openapi-spec

We drop:
* mergify-import-check-worker
* mergify-import-check-web
* mergify-import-check-db-update

This is the first PR to do this.

Change-Id: I8609b87c507edd2cecd511d1b4cd6f27741d4a76