refactor(event-forwarder): put forwarded events in Redis (#7187)
Since we put forwarded events list in memory, we have some warnings about
high memory usage of web worker, worker may even be restart.
And this also increases the chance to loose events.

This change put the events to forward in Redis and a new service drain
this list to send them to shadow-office.

I already enabled in production the new service

Related INC-80