refactor: factorize retry handling of http requests (#7177)
This change replaces the repetives code around tenacity.AsyncRetrying by
a httpx custom extension.

As our httpx transport already a retry mechanism, this changes directly
pass to the http client request the retry predicate and logging message.

The transport retry mechanism will use that on top of it's actually
configuration.