fix: invalid bot account credentials (followup) (#7243)
Currently, we retry forever when a bot_account credentials is invalid.

As this requires the user to relogin we should report it correctly and
not retry.

This is the same we done for edit action but for all other places we use
a bot_account with an HTTP request.

Related MRGFY-2503

Change-Id: I82217ec50d7b6413db14c8e52ad683a9bd8b1444