fix: canceled pr that are implicitly merged (#7261)
When a pr (1) has a commit A and another one (2) has a commit A and B. If (2) is merged by us before (1), (1) will be implicitly merged by (2). Github will say that is us that merge (1) because it's us that merge commit A through (2). We have to distinguish those two cases because if the engine detects that a pr is merged by us it will try to remove it from the train but in this case (1) is not in any train because it's not us which merged it.

That pr add the distinguish code

Fixes: MRGFY-2504