fix(queue): Build unqueue msg with queue name from action config (#7314)
When using commands and a pull request was unqueued because of a failing check, the queue command would try to find the queue name from the convoy but would not be able to find the pull request due to it already being unqueued. Since we already have the queue name in the action config, we can use that instead which is much easier and faster.