fix(queue): ensure final draft pr state is really final (#7306)
In case of check_mergeability was called from the draft pr and a CI move
back to pending, this may mess up the merge-queue as the PR got closed
and may interrupt this buggy CI blocking the queue forever. Then the
customer don't understand why we closed the PRs not merged.

This change ensures we really don't update the state when the state in
not pending.

Fixes MRGFY-2539

Change-Id: I4ef7363ec04a161f0a4e8e75d3d375d257d0a3a7