feat: add tenacity retry on integrity error (#7315)
Since we have more and more usage of database through different process, we begin to have integretyerror especially on GitHubAccount or GitHubRepository because of parallel sessions. Added tenacity retry on session to retry once when it occurs.

refactor ci_event_processing a bit to move the session closer to where it s used.

Fixes: MRGFY-2533