ci(create-linear-docs): fix stateId (#7372)
The workflow still fails with this:

{"errors":[{"message":"Variable \"$stateId\" is not defined.","locations":[{"line":1,"column":201},{"line":1,"column":1}],"extensions":{"code":"GRAPHQL_VALIDATION_FAILED","type":"graphql error","userError":true,"meta":{}}}]}

It does not make sense since stateId is passed as a GraphQL variable so
it is available. For whatever reason, the GraphQL Linear API does not
see it while it's passed by the jq output JSON correctly.

I tried locally to make this work and the best solution I found is to
just harcode the stateId in the query which should be fine.