fix(rules): Fix queue conditions evaluation on QueuePullRequest (#7384)
When the TrainCar asks the pull requests to be evaluated, it pass a list of `QueuePullRequest` to the evaluator and the evaluator then evaluate both the `merge_conditions` and `queue_conditions` on those. This caused the `queue_conditions` to be evaluated on the draft pull request, when there was one, instead of the actual queued pull requests.

Related MRGFY-2340