feat: Add DbPopulator loader and populated_db fixture (#7362)
Add DbPopulator loader and populated_db fixture. 

In the future it would be great to create data needed for tests here to increased number and varity of data and scenario. It could be great to convert all the datacreation that we have now in this way for the same reason. We could reuse the same basic scenario when we only need some basic data.

Fixes: MRGFY-2490