fix(subscription): don't slow down engine when Shadow Office is down (#7388)
The Shadow Office is used by the engine to get the subscription details.
These details are cached for an hour. When the cache expires and the
Shadow Office is down, the engine retry to get the subscription details
on every run, thus slowing down the process.

We could either reduce the HTTP request timeout or number of retries,
which is already done (see mergify_engine/clients/dashboard.py). Or we
could renew the validity of the cache for an hour and avoid to make too
much HTTP requests that times out.

Fixes MRGFY-1914