fix(date): Fix comparison of `UncertainDate.tzinfo` objects between `zoneinfo` cache clears (#7462)
We have a test that clears the `zoneinfo` cache
(`zoneinfo.ZoneInfo.clear_caches()`) to test the comparison of two `Schedule` objects between a cache clear. This caused the comparison of the `UncertainDate.tzinfo` attributes to fail when ran in the same process as the cache clear.

This commit also reverts a part of873a115eb5dc1a33dbef920720876bb44c8367c2 and remove the test retries of `test_parse`.