fix: Search failed step on conclusion different from success (#7473)
Failed steps were searched only on conclusion equal to `failure`, but in this case, the failed step had conclusion `canceled`. I also had a raise at the end of the loop to raise in an early stage when the failed step is not found.
Plus, I fixed the raise in `download_failed_step_log` because it was missing a `raise`.

Fixes: MRGFY-2585

To fix the broken job in the database, we have to execute this query after merging this PR:

```sql
UPDATE
    gha_workflow_job
SET
    failed_step_number=5,
    failed_step_name='Build 🔧 & Test 🔍'
WHERE
    conclusion = 'failure'
    AND name = 'record (GitHub Enterprise Server, FIXTURES_RECORDER_CONFIG_GHES_DEV)';
```