fix(ci-processing): don't retry job with missing steps (#7578)
The ci-processing has been dead since the first time we got this Sentry. A ton of CI events are blocked in Redis.

If a job doesn't have expected steps, we now just log it once and
store it as-is. The log-embedder has been fixed to ignore such job already.

We still log it in Sentry when this occurs, but we stop retrying it.

This will avoid spamming Sentry and Datadog we same error in loop.

Fixes ENGINE-3EQ
Fixes MRGFY-2588

Change-Id: I4ea82410134e24943e2f362f3e412613017ca2f3