fix(log-embedder): handle 410 from GitHub (#7600)
This changes adds a new enum to track if we successfully get the log and
embedded it.

This permits to stop processing job where the log is gone and (in the
future) where too many errors occurs when trying to download it.

This also adds a new columns constraints to ensure any insertion failed
if the three linked columns are not in sync.

Fixes ENGINE-3EX
Fixes MRGFY-2599

Change-Id: If16c48a0a626b5808d15a0a66e645e2191c3fadb