fix(log-embedded): improve zip file finder (#7637)
job id naming rule is:
* must start with a letter or _ and contain only alphanumeric characters, -, or _

But when the job contains a matrix names of the matrix items are added
between parenthesis. And this part does not follow the same rules.
At first glance, the rule is the same that we found previous windows
incompatible chars must be removed.

We need to check the max length of the directory names and how string
are truncated when the matrixes items contains unicode chars with more
than 2 bytes.

Partial fixes: ENGINE-3F2
Fixes MRGFY-2605