fix(log-embedded): continue when we failed to embed log (#7654)
As it will be hard to always find the log file in the zip for example

This changes replaces the RuntimeError by a LOG.error to allow
processing of next jobs.

And add a generic Exception catch to continue to proceed next jobs.

The failing jobs will be retried one minute later.

We still have an issue when the number of job is >=
LOG_EMBEDDER_JOBS_BATCH_SIZE as the SQL select will pick only
non-working jobs.

To solve that we need to continue to improve the error handling by
retrying this case later with a backoff.

Depends-On: #7601

Change-Id: I359a0f49cc050e28d845e1389a945e5bf19cf9ae