fix(log-embedder): don't use as_dict() for logging (#7689)
as_dict() is far too verbose as it dumps all attributes and all
relashionships attributes.

Our logging sends data to Sentry and Datadog we must not go over
a log line as a limited size. For example Datadog is 1Mo

Going over means we lost logs and may not see some Sentry.

Change-Id: I7832d6fde4842aafe5bd4ddd2c8bc351a26641b7