fix(log-embedded): ensure log_embedded is not too big (#7674)
The previous code just pick the same lines we uses for getting the
embedding but not cleaned.

This clean may removed a lot of token in one line, like a failure during
a `git pull` will print ton chars in one line to show the progress.

Also the non cleaned logs has `\n` char and it one more token.

This change put a limit on the number of tokens log_embedded to ensure
we will always fit into the chat model.

Fixes ENGINE-3F5

Change-Id: I21b74047034cded1c832fd89fc06d4e4b8519861