test(run_full_engine): make it no flaky (#7713)
From time to time, tests failed because delayed refreshes are not send,
then the engine didn't run because no refresh are in the todo list.

This is because the run_full_engine loop stop everything before the
refresh are emitted.

This changes checks if delayed_refresh must be emitted before stopping
workers.

This ensures the engine will get the refresh events that trigger it.

This solves the flakyness of test_time_to_merge_endpoint_with_schedule
and maybe many others using run_full_engine()

Change-Id: I2c7bbd2fdcdfb14d674077738eaac4d60eb90ecf